Successfully Linked FreeType

+++ include/display_text.hpp
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <string>
 #include <ft2build.h>
+#include <freetype/freetype.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
@@ -14,7 +15,7 @@ struct Character
     unsigned int TextureID; // ID handle of the glyph texture
     glm::ivec2 Size;        // Size of glyph
     glm::ivec2 Bearing;     // Offset from baseline to left/top of glyph
-    unsigned int Advance;   // Offset to advance to next glyph
+    long Advance;   // Offset to advance to next glyph
 };
 
 class DisplayText
